requirements.txt
aiohttp==3.6.2
APScheduler==3.5.1
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
datetime==4.3
discord.py==1.3.3
emoji==0.5.0
idna==2.10
idna-ssl==1.1.0
multidict==4.7.6
pytz==2018.3
requests==2.24.0
six==1.11.0
typing-extensions==3.7.4.2
tzlocal==1.5.1
websockets==8.1
yarl==1.4.2
